Change the cutoff for considering old aircraft
Aircraft not seen for 30 seconds are ignored when parsing aircraft.json
adsb-receiver project used 15 seconds but that is a little harsh for
cases where the aircraft is not seen for long.
Using a longer timeout is also problematic as aircraft no longer seen
are still counted.
Anyway this is not an exact science.
For more accurate stats, parsing more often than every 60 seconds would
be required, but i'm not sure how to implement that.