Handle non number values at file creation
The data stored for local messages differs between readsb and dump1090-fa when an external receiver such as airspy is used. dump1090-fa records message rates in dump1090_messages-local_accepted.rrd as numeric 0, whereas with readsb it is stored as nan. This is problematic for gnuplot which ignores lines with nan fields completely. Changing it here keeps the data consistent whether dump1090 or readsb are used.